Minor clenaup
